Consolidate AWS DSL enum spelling
Promote carina-aws-types::dsl_enum_value as the canonical implementation, route AWS codegen through it, and normalize slash/colon enum separators for DSL aliases.

Fixes carina#3463.
