+++ carina-aws-types/Cargo.toml
@@ -15,3 +15,4 @@ indexmap = "2"
 
 [dependencies]
 carina-core = { git = "https://github.com/carina-rs/carina", rev = "0a70d5bfd213d50e1bc5bca22c964f420e51c401" }
+heck = "0.5"
