+++ carina-codegen-aws/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod dsl;
 pub mod resource_defs;
